+++ Tests/demistomock/demistomock.py
@@ -1,3 +1,5 @@
+from __future__ import print_function
+
 import json
 import logging
 import uuid
@@ -429,17 +431,17 @@ def setLastRun(obj):
     return None
 
 
-def info(*args):
-    log(args)
+def info(msg, *args):
+    logging.getLogger().info(msg, *args)
 
 
-def error(*args):
+def error(msg, *args):
     # print to stdout so pytest fail if not mocked
-    print(args)
+    print(msg, *args)
 
 
-def debug(*args):
-    log(args)
+def debug(msg, *args):
+    logging.getLogger().info(msg, *args)
 
 
 def getAllSupportedCommands():
@@ -553,7 +555,10 @@ def handleEntitlementForUser(incidentID, guid, email, content, taskID=""):
     return {}
 
 def demistoVersion():
-    return {}
+    return {
+        'version': '5.5.0',
+        'buildNumber': '12345'
+    }
 
 def integrationInstance():
     return ""