+++ Packs/ExportIndicators/Integrations/ExportIndicators/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+  - Removed `Long Running Instance` from instance configuration.
 
 ## [20.4.0] - 2020-04-14
   - Fixed an issue where running **On-Demand** mode an error appeared if export was not initialized.