+++ Packs/ExportIndicators/Integrations/ExportIndicators/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-  - Removed default initial value for **Listen Port**.
+Removed the default initial value for the **Listen Port** port.
 
 ## [20.4.0] - 2020-04-14
   - Fixed an issue where running **On-Demand** mode an error appeared if export was not initialized.