+++ Packs/ExportIndicators/Integrations/ExportIndicators/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [20.4.0] - 2020-04-14
   - Fixed an issue where running **On-Demand** mode an error appeared if export was not initialized.
   - Now when the *query* argument in the ***eis-update*** command is not supplied, the query from the integration parameters is used.
   - Added a feature to output *csv* and *XSOAR-csv* formats as textual web pages. This can be done by: